+++ CHANGES.rst
@@ -6,6 +6,11 @@ Async PRAW follows `semantic versioning <http://semver.org/>`_.
 Unreleased
 ----------
 
+**Added**
+
+- :meth:`~.WikiPage.discussions` to obtain site-wide link submissions that link to the
+  WikiPage.
+- :meth:`.revert` to revert a WikiPage to a specified revision.
 - The configuration setting ``refresh_token`` has been added back. See
   https://www.reddit.com/r/redditdev/comments/olk5e6/followup_oauth2_api_changes_regarding_refresh/
   for more info.